Run update-index as a single background process
Reorganize the transfer queue to provide a channel to watch for object
OIDs as they finish. This can be used in the `get` command to feed a
goroutine that will copy the file to the working directory and inform
the update-index process about it as the transfers finish. This leads to
a greatly reduced amount of time spent updating the index after a get.


Former-commit-id: ebc81aedb9af1dfb24f492ff93115000c2da18ea