Fix annoying keychain errors when running integration tests on OS X
This was caused by the fact that you can't disable the osxkeychain helper
if it's set at the system gitconfig (which is the default in recent git),
and the git credential 'store' operation will call ALL helpers, not just
the special one created for the integration tests. Because the test user
environment doesn't have a keychain this caused lots of annoying popup
error messages saying "A keychain could not be found to store X".


Former-commit-id: 332b22944ec7a544e476b2062b4e0471fdbf31d3