Don't use make here
When using make with append, both a capacity and length must be used in
the make call. If only the length is used the array will have len zero
values at the beginning and append will start adding values at the end.
This is a small number of values so just skip make altogether.


Former-commit-id: 711f483767083627d729845f5550249110aa2eca