Add optional remote arg to fetch & pull, and default to tracking remote
The first arg to fetch & pull is now a remote. In addition, the default
remote if you don't specify is now the tracking remote as in `git pull`
if it exists, and origin if that's not set. This makes it more consistent
with the underlying git workflow especially in triangular fork setups.


Former-commit-id: b5df63c071d22ed768e871e3557ad6cb4a89c895