Change RecentRefs to RecentBranches & optimise so we can get date in 1 cmd
Previously were trying to support tags too, but this is too expensive.
There's no commit date on a tag object so you can't interleave them in
the list effectively, would have to use %(taggerdate) in a second call
which is non-optimal but also wrong (could tag retrospectively). Only way
to get tags on the same basis as branches is to query each tag's
dereferenced commit separately which won't scale. Users can always
fetch tags on the command line if they want anyway.

Doing this also lets us get the commitdate as part of the for-each-ref
which is much more efficient.


Former-commit-id: 1538c8920cfd69e9bf704c47f1a1b687162af0c5