Test disabled batch support more locally not globally on the server
The test for lack of batch support turned it off globally on the test
server then turned it back on, but since the server is shared between
parallel tests, this introduces a race condition which can break other
batch tests randomly. Instead, use a specific repo name to flag it.


Former-commit-id: d4f63584c3ba6811bbd9f33a64b38c16dff61b47