smudge --skip will now smudge if object is in cache.
This change makes it so --skip does the smudge if the file
is in cache. It still skips automatic downloads as one would
expect.

This is more convenient, since it prevents one from having
to 'git lfs pull' after every 'git checkout' if the needed
objects are already available in .git/lfs/objects.

Note that the --skip option for smudge is documented as
follows:
"Skip automatic downloading of objects on clone or pull."
So even after this change we're still in line with the
documentation.

Former-commit-id: 71bfd4e405681011c6bb827dc8b7897a1dfdb965