Fix fallthrough when git-lfs-authenticate returns an error
If using an SSH URL and git-lfs-authenticate returns an error, without
this change the error is just reported to the trace but execution
continues.

A common result is a completely made up LFS URL being returned which
then 404's, which ends up setting lfs.batch=false. If the server doesn't
implement the non-batch interface, this config then breaks LFS even
after the SSH issue is resolved & has to be manually removed.


Former-commit-id: a79cb1c84da69423ee793f98965838d15b5a0992