Explain a subtlety of client handling for "download" batch action
The LFS client uses the presence of the "download" action on an "upload" request to deduce that the object content is already present on the server: the object content must be present, or the server could not offer a "download" action. Since the server already knows the object content, the client does not need to upload a new copy and can skip the transfer step.

Make this interaction between the client and server explicit in the documentation.

Former-commit-id: 5e2e61cd8730c88c28bdee9bbfd59adddef1f329