travis-ci: add a build job to test against upcoming versions of Git
The Git source version is defined by the variables GIT_SOURCE_REPO and
GIT_SOURCE_BRANCH (which can also be a Git hash). The Git source is
checked out, compiled, and installed on the Travis CI Linux machine.
Afterwards all Git-LFS tests are executed against the this Git version.

This setup helps to test features that depend on changes in the
development branches of Git and Git-LFS.


Former-commit-id: e5756d3d22500b4d223092a85f04c0f2559d17ba