Bugfix: cannot push to scp style URL
When push to a scp style url, like:

    git push git@github.com:git-lfs/git-lfs.git HEAD

Git-lfs will complain about bad url.  This is because golang 1.8 has
changed the implement of `url.Parse` method.  It will fail with a nil
pointer and an error if the remote url is a scp style URL.

Make a golang compatible test, I.E. check null pointer, then check u.Scheme.

Signed-off-by: Jiang Xin <xin.jiang@huawei.com>

Former-commit-id: 760a4517a0e89f5ad2b83a7442071bb0f4a36ebe