fix multiple `git update-index` invocations
The gitIndexer starts a new `git update-index` process when the first
file is given to it. A reference to this process is stored in the `cmd`
variable of the gitIndexer struct. This variable was always `nil` after
the refactoring in ae4d2611 ("move `git update-index` invocation to git
package", 2017-08-02). Consequently, a new `git update-index` process
was invoked for every file.

Fix this by refactoring the `git update-index` invocation, again. This
time we return the subprocess pointer. This is necessary because we
want to call `Wait()` on the process when the gitIndexer is closed.


Former-commit-id: e176f361d6263fde2574a24b09d145e2370be4ed