add basic example in main manpage
Every time I go back to git LFS after a long break, I wonder how it
works and how to get started. The first manual page is currently a
reference to a fairly large number of other manual pages and, as the
core documentation of LFS, does not really help users in getting basic
things done.

In comparison, the LFS homepage has a good primer, so I have taken it
as an example for a basic example that would work in the manual page
to help users get out of the weeds quickly.


Former-commit-id: c6383b9ca36c29b6300828da48b532781c9d01f9