Fix cobra help command to work properly with newer version of cobra
This was initially broken upstream in
https://github.com/spf13/cobra/commit/d6bf4ef243b7f3b96d8a4249e63af711ef33d53e.
They tried to fix it with
https://github.com/spf13/cobra/commit/ded646f9783b3b64db74703f1bfc9d02f683ccc7
but functionality was not actually restored.  This takes the help
command code from
https://github.com/spf13/cobra/commit/ded646f9783b3b64db74703f1bfc9d02f683ccc7
and explicitly sets it as the help command.  On one hand this is a bug
that should arguably be fixed upstream, but on the other hand I believe
this is now the way they intend people to use the `help` functionality.

Former-commit-id: 892907ae1cbe482f08266f23485692ff47cc75d1