Makefile: list prerequisites for resource.syso
When building Git LFS v2.5.0, we noticed that the resource.syso was
unchanged between builds and thus retained the old version information
found in 'master' before v2.5.0.

Hence we built binaries in bin/git-lfs.exe with a commit tagged
'v2.5.0', but a version info indicating that the binary was of the
v2.4.0-era.

Let's list out the dependencies of 'resource.syso' explicitly, and thus
rebuild it when any of those change. In particular, it is unlikely the
installer icons. But, if we change the version information during a
release in versioninfo.json, we should update the resource.syso file
accordingly.


Former-commit-id: ab18b1281ad139d07dd349420fb95c90b56fe6e2