tq/basic_upload.go: disable Content-Type detection by configuration
For the majority of Git LFS's lifetime, Git LFS objects have been
uploaded via HTTP and contained the header:

    Content-Type: application/octet-stream

In [1], this changed and we began attempting to guess an appropriate
Content-Type header based on the first 512 bytes of the file being
uploaded.

This breaks some hosting platforms' implementation of this API, so let's
offer a way to disable this guessing to allow users to upload objects
again on affected implementations.

[1]: a1736293 (Merge pull request #3137 from
     calavera/set_upload_content_type, 2018-07-23)


Former-commit-id: 938df5c47f343279e8ece0aa5f86e57be4dbdf11