Makefile: hide the 'go list' invocation
For the same reasons as [1], let's hide the invocation of 'go list',
too.

This rule is used to make sure that we don't have any un-vendored
third-party dependencies, but the 'go list' pipeline is awkward,
lengthy, and (like the 'goimports' one), doesn't often carry useful
information.

Again, if one such dependency does exist, than this pipeline will ensure
that it is propagated outward. But, if not, let's make sure that the
original pipeline is hidden to shorten the build output.

[1]: 3c0c8ac3 (Makefile: silence GOIMPORTS invocation, 2018-07-31)


Former-commit-id: 1f6d3a7f62384e800c3d7705cd49b84b96018c63