Makefile: un-qualify path to 'commands/mancontent_gen.go'
Because this Makefile is located in the repository root, all paths
listed in this Makefile are relative to the root of the repository as
well.

Hence, there is no need to fully-qualify the path from $(GOPATH) down to
commands/mancontent_gen.go, and the path can be specified as the later
part of the previous clause.


Former-commit-id: 023d5ba6ffd898da28fca4717c6646456bde1c63