tools/filetools.go: extract fastWalker consumer func
When calling FastWalkGitRepo, the implementation constructs an
instance of the *fastWalker type, and then consumes items from the
<-chan until it receives a close().

This was left mostly unchanged since 3b9e629d (use a single channel,
2016-11-29), when we had only the single function `FastWalkGitRepo`.

In a subsequent commit, we will introduce a new function
`FastWalkGitRepoAll` with does _not_ exclude patterns found in a
repository's .gitignore(s).

In preparation for this function, which we assume to share the
implementation of `FastWalkGitRepo` in its entirety (modulo the
construction of a `*fastWalker`), let's extract this fastWalkCallback
function to avoid repeating ourselves.


Former-commit-id: c9cde1189bcbde5a7df2760b66eabe72d0ed0441