tools: add function to perform shell quoting
In certain cases, such as the GIT_SSH_COMMAND environment variable,
we'll need to pass data to the shell.  Since some of our arguments
contain spaces, we'll need to quote those values.  Introduce a function
to quote strings so that they're suitable for the shell.  For
simplicity's sake, assume anything beyond identifier characters, dot, at
sign, and hyphen-minus needs to be quoted and use single quotes for
doing so.  Handle single quotes in a string in the same way as Git does.


Former-commit-id: 69b534ece0d032b7005c16129f41caacfaaf0ee3