lfsapi: pass GIT_SSH_COMMAND to the shell
Git documents GIT_SSH_COMMAND as being passed to the shell.  However, we
were not doing that, leading to unexpected behavior.  Determine if we're
getting the SSH command from GIT_SSH_COMMAND and if so, handle things
specially.

Set a flag indicating that we need to pass data to the shell, and return
this flag as well as the standard data.  Add a function that formats
arguments appropriate for either the shell or non-shell case.  Note that
we cannot do the shell formatting in sshGetExeAndArgs, since we want to
append arguments to the command line and the way we do that differs
between the shell and non-shell variants.  Continue to parse the first
part of the shell command so we know if we're using plink or
tortoiseplink.

Update the tests to use both sshFormatArgs and sshGetExeAndArgs
together, since it no longer makes sense to test them independently of
each other.


Former-commit-id: 7990ab17a4ec9b039df61f3e08873f2c687379e5