Makefile: fail on improperly vendored dependencies
Since [1], we have caused 'make lint' (or any recipes that list 'lint'
as a prerequisite) to fail if Git LFS imports non-standard packages
outside of its vendor tree.

However, this approach is deficient in that it does not correctly report
whether there are improperly vendored dependencies, since the pipeline
ends with "|| exit 0".

To prepare for a change in the invocation of "go list", let's make sure
that a final "grep ." fails (i.e., that the output of the pipeline up
until that point is empty) by adding a final "grep ." and negating the
exit code of the entire pipeline.

[1]: a1305a8c (Makefile: replace script/{fmt,lint} with '{fmt,lint}'
target, 2018-07-16)

Co-authored-by: brian m. carlson <bk2204@github.com>

Former-commit-id: 1819e679bbf2e5cb8296bac6a67ce5502d1ba00c