commands/command_ls_files.go: don't accept "git lfs ls-files -- --all"
In 05d65e75 (commands/ls-files: do not accept '--all' after '--',
2018-03-20), we encountered an interesting situation in which the
following command invocation:

    $ git lfs ls-files -- --all

Was treated as if "--all" were a reference name, and thusly passed to
git.ResolveRef(), and ultimately to "git rev-parse", as "git rev-parse
--all", which is a valid execution of "git rev-parse".

Thus, we'll get an odd result. Instead, let's trap that case early, and
return an appropriate error.


Former-commit-id: 6db67f2ca50bf0a4fc886f0c17f69b34dad72779