go.mod: use latest github.com/olekukonko/ts
We use package github.com/olekukonko/ts to issue IOCTL's to gather the
terminal size of the calling terminal in order to determine the length
of lines we attempt to print.

Until now, we have used revision ecf753e7c962, which is over 4 years
old. Since then, github.com/olekukonko/ts has introduced support for
Solaris, which we previously could not build or run on because package
syscall on Solaris does not export an IOCTL constant.

Let's upgrade to the latest version, and bring in support for Solaris.


Former-commit-id: d32ff9b429cff488a72cd71b8119617f68e10084