t/Makefile: mark test sources as .PHONY
We have used 'make' in the testing directory 't' to help run integration
tests, including starting and stopping a Git LFS test server, and
environmental setup bits.

We do so by writing a recipe that matches all testing files that we'd
want to run, and then running commands based on which one we picked.

Since we don't generate a file that was missing when running one of
these sources, 'make' will say that the file is up-to-date, unless we
run with 'make -B' (to force a re-make).

This is annoying to type, and can be avoided by marking the test targets
as .PHONY, indicating that they don't produce a file, and are instead
synthetic targets that can be run repeatedly without `make`'s up-to-date
checking.


Former-commit-id: f8147219fda9278c6f3697a72a1738862105fa31