lfsapi: pass Git config info when finding SSH command
In a future commit, we're going to want to read from the Git
configuration file so that we can use core.sshCommand as a possible
place to find our SSH client.  In preparation, pass the Git
configuration information down into the functions that do SSH command
lookup can have it available.

This commit should result in no functional change.


Former-commit-id: fcefe6890dbdf81facfcf8df7e34121c6ce02877