subprocess: add a function to shell quote a single string
We currently have a function to shell quote a list of strings, but in
some cases we may want to just quote a single string.  To avoid the
complexity of having to create a bunch of temporaries and then index out
the result, create a function that shell quotes just a single string.


Former-commit-id: 9580353dca4d4a46dff902318a6322e1fb3564c2